And fix the deploy script so that it not only deploys ue4_5, but it will detect future ue4_* targets.

+++ script/deploy.lua
@@ -23,6 +23,7 @@
 -- This is not a good solution and will be fixed when it becomes more important.
 
 require "script/lib/util"
+require "script/lib/platforms"
 
 local version = io.open("version", "rb"):read("*line")
 local major, minor, patch, update = version:match("^(%d+)%.(%d+)%.(%d+)-(%d+)-")
@@ -38,18 +39,20 @@ local uev = ((major * 100 + minor) * 100 + patch) * 1000 + update
 os.execute("rm -rf Den* script version TODO")
 
 -- create ue4 deployments
-for _, v in pairs{"ue4_3", "ue4_4"} do
-  os.execute(("mkdir -p plugin/%s"):format(v))
-  os.execute(("cp -r ue4/Plugins plugin/%s"):format(v))
-  os.execute(("mkdir -p plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps/deps"):format(v))
-  os.execute(("mkdir -p plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps/lib"):format(v))
-  os.execute(("cp -r include plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps"):format(v))
-  os.execute(("mv lib/%s plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps/lib"):format(v, v))
-  os.execute(("rm plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps/lib/%s/x32/frames_renderer*"):format(v, v))
-  os.execute(("rm plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps/lib/%s/x64/frames_renderer*"):format(v, v))
-  os.execute(("cp -r deps/boost_1_55_0 deps/jpeg-9 plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps/deps"):format(v))
-  os.execute(("sed -i s/DEVELOPMENT/%s/ plugin/%s/Plugins/Frames/Frames.uplugin"):format(version, v))
-  os.execute(("sed -i 's/: 1,/: %d,/' plugin/%s/Plugins/Frames/Frames.uplugin"):format(uev, v))
+for v in pairs(projects) do
+  if v:find("ue4") then
+    os.execute(("mkdir -p plugin/%s"):format(v))
+    os.execute(("cp -r ue4/Plugins plugin/%s"):format(v))
+    os.execute(("mkdir -p plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps/deps"):format(v))
+    os.execute(("mkdir -p plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps/lib"):format(v))
+    os.execute(("cp -r include plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps"):format(v))
+    os.execute(("mv lib/%s plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps/lib"):format(v, v))
+    os.execute(("rm plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps/lib/%s/x32/frames_renderer*"):format(v, v))
+    os.execute(("rm plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps/lib/%s/x64/frames_renderer*"):format(v, v))
+    os.execute(("cp -r deps/boost_1_55_0 deps/jpeg-9 plugin/%s/Plugins/Frames/Source/ThirdParty/FramesDeps/deps"):format(v))
+    os.execute(("sed -i s/DEVELOPMENT/%s/ plugin/%s/Plugins/Frames/Frames.uplugin"):format(version, v))
+    os.execute(("sed -i 's/: 1,/: %d,/' plugin/%s/Plugins/Frames/Frames.uplugin"):format(uev, v))
+  end
 end
 
 os.execute(("mkdir frames-%s && mv * frames-%s"):format(version, version))
